\name{gitProjectExtras-package}
\alias{gitProjectExtras-package}
\alias{gitProjectExtras}
\docType{package}
\title{Extra Functions for Integrating with git based projects}
\description{Allows linking code and outputs with information about git revision numbers and storage 
             locations. Also contains tools for setting up a proper project structure.}
\details{
\tabular{ll}{
Package: \tab gitProjectExtras\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-15\cr
License: \tab GPL 2\cr
}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{Douglas Kelley \email{douglas.kelley@mq.edu.au}}
%%\references{
%%~~ Literature or other references for background information ~~}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
\keyword{ package }
%%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
%\examples{

%}
