\name{match.call.string}
\alias{match.call.string}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Match Call String}
\description{Returns the name of the current function}
\usage{
match.call.string(n = 1, call = sys.call(sys.parent(n = n)), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{call position used by \code{\link{match.call}}. Default \code{n = 1} returns the function anme from which \code{match.call.string} is called. \code{ n} > 1 returns the function name \code{n-1} up the call tree}
  \item{call}{an unevaluated call to the function specified by definition, as generated by call and used by \code{\link{match.call}}}
  \item{\dots}{Additional arguments passed to \code{\link{match.call}}}
}
%%\details{}
\value{function name from which match.call.string is called. Returns "N/A" if not called within a function or if n is greater than the call tree}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{Douglas Kelley \email{douglas.kelley@mq.edu.au}}
%\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
    \code{\link{match.call}}
}
\examples{
match.call.string()
yayandwow <- function()
    match.call.string()

yayandwow()

#########################

yay <- function(...) {
	wow <- function(...) match.call.string(...)
	wow(...)
}

yay()
yay(n=1)
yay(n=2)
yay(n=3)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ call }
\keyword{ name }
